[refactor] moved routes to respective models
