+++ backend/tests/test_purchase_routes.py
@@ -5,13 +5,14 @@ def test_vending_machine_records(client: FlaskClient) -> None:
     """Test machine records."""
     response = client.get("/item/records")
     assert response.status_code == 200
-    json = response.json
-    assert json["status"]
+    response_json = response.json
+    assert response_json["status"]
 
 
 def test_product_records(client: FlaskClient) -> None:
     """Test item records."""
     response = client.get("/machine/records")
     assert response.status_code == 200
-    json = response.json
-    assert json["status"]
+    response_json = response.json
+    assert response_json["status"]
+
